scripts/requirements.txt
black==26.1.0
click==8.3.1
et_xmlfile==2.0.0
mypy_extensions==1.1.0
openpyxl==3.1.5
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.2
pytokens==0.4.1
ruff==0.15.2
